Prevent tab borders from colliding with parent if show_border=false (Fixes #228)
The core idea of show_border = false is to get rid of the border around
the widget and set the margins to zero to visually embed the widget
directly inside the containing widget, without introducing additional
borders.

To facilitate such design choices by applications, Greybird's tabs need
to be slightly moved to the right to prevent them from colliding with
the containing widget's borders.

A screenshot of how this looks in Greybird and other themes is supplied
in #228.